php: ajax-handlers transferred from functions.php to ajax.php
